Fix test

+++ tests/test_models/test_place.py
@@ -24,7 +24,7 @@ class TestPlace(unittest.TestCase):
         self.assertIsInstance(Place.number_rooms, int)
         self.assertIsInstance(Place.number_bathrooms, int)
         self.assertIsInstance(Place.max_guest, int)
-        self.assertIsInstance(Place.price_by_night, float)
+        self.assertIsInstance(Place.price_by_night, int)
         self.assertIsInstance(Place.price_by_night, int)
         self.assertIsInstance(Place.latitude, float)
         self.assertIsInstance(Place.longitude, float)
